Provision celery on app and worker.
We need celery running on both the worker machines but also on the app server.
The worker machines will have a configurable number of workers and threads per
worker so that we can maximize our server usage. On these workers celery runs as
a service but will have access to the rest of the application code though django
will not be served via Nginx. In dev this will spin up a new worker vm that will
connect to the redis service on the service VM.