Stub celery workflow with task runners.
This prepares a chained workflow of tasks that are run by workers on one or more
worker servers. We POST the input data to an API endpoint which feeds the the
input data to a task that will kick of the GeoTrellis job. Right now this just
makes an HTTP request to fetch azavea.com so we can see real network activity.
The results of this are then fed to the model runner. Results of the model are
stored in the database. A uuid is returned to the client upon hitting the API
endpoint. This uuid can be used to check the status of the job.