Create a working stub tiler and tiler VM.
We want to serve tiles for our national scale data such as congressional
district from Windshaft which is a nodejs tiler that wraps calls to mapnik. The
tiler will live on its own VM for development and will be one of potentially
many machines on production. Windshaft serves on port 4000 but this is proxied
through Nginx on port 80. The tiler grabs info from the database on the services
machines and uses Redis on the services machines. In this stub, we only serve
congressional districts.