Keep map model in sync with map view location
* Update the map model lat, lng, and zoom attributes
when the map is panned or zoomed.
* Change the scope of attributes that trigger MapView.updateView.
This has the downside of triggering an event when none of the
location related attributes change, but the upside of only
triggering once when all of the attributes change. Now that we
are keeping the map model up-to-date with map position, the
extra calls to update the view should have no effect. Doing
it this way also fixes an issue similar to Leaflet/Leaflet#2485
that would happen if you searched for and selected two
locations that were nearby (ex. 1234 Market St and City Hall)
one after the other.
* Add tests for map model/view position updating.

Fixes #72