Create templates for `/model` page.
This patch plucks portions from the static version of the `/models` page
and turns them into templates.  The controller, the views, and the
models are all based on those for `/analyze`.

In addition, the map is now visible on the `/models` page.

Connects to azavea/model-my-watershed#130
Connects to azavea/model-my-watershed#131