Keep analysis.
This patch causes the application to keep a copies of the
`AnalyzeTaskModel` objects correspondent to each area of interest that
it knows about, instead of recreating them all of the time.  This has
the effect of removing the pause that occurred when one left and then
returned to the `/analyze` page without changing the area of interest.

Connects to azavea/model-my-watershed#109
Connects to azavea/model-my-watershed#145