Create ItsiUser model and migrations
We need to store the ITSI User ID of the user somewhere. Instead of
extending the core auth_user model / table, whose structure other apps
may assume, we create a new user_itsiuser model / table, which has a
foreign key link to our internal user, and a single itsi_id field for
storing the ITSI User ID.

We may add more fields to this model in the future.