Predefined shape outlines.
Instead of giving a selection of predefined shapes in the dropdown, give
a selection of types of shapes (congressional districts, hucs, &c).
When a type is chosen, the outlines of those shapes will appear on the
map, allowing the user choose that way.

This patch assumes the existence of two environment variables on the
`app` VM: `MMW_TILER_HOST` and `MMW_TILER_PORT`.  If they do not exist,
the corresponding variables within Django are given the defaults
`localhost` and `4000`, respectively.

Also, the name of the column containing the polygon in the
modeling_district table has been renamed from `polygon` to `geom`.

Also, `src/mmw/mmw/settings/base.py`, `src/mmw/apps/modeling/views.py`
and `src/tiler/server.js` have all been updated to support adding new
geometry-bearing tables more easily.

Also, If the user clicks on e.g. 'Draw Area' while the outline layer is
enabled, it is cleared.  If the user then decides not to draw a polygon,
the outline layer does not return (this is intentional).

Connects to #176
Connects to #210