Predefined shape outlines.
Instead of giving a selection of predefined shapes in the dropdown,
instead give a selection of types of these shapes.  When a type is
chosen, the outlines of those shapes will appear on the map, allowing
the user choose that way.

This patch assumes the existence of two environment variables on the
`app` VM: `MMW_TILER_HOST` and `MMW_TILER_PORT`.  If they do not exist,
the corresponding variables within Django are given the defaults
`localhost` and `4000`, respectively.

Connects to #176