Add data
This is taken from the CSV file that powered the original flash version.

Some data is missing, since we added 4 new land types: Chaparral, Tall Grass
Prairie, Short Grass Prairie, and Desert.

Also, some data does not add up. For example, when Precipitation is 2, Land is
Low Intensity Residential, and Soil is Sand, ET is 0, R is 0.4 and I is 1.5
which adds up to 1.9, not 2, thus there is empty space in the bottom of the
water column.

The `rainKey` variable translates the slider positions of 0-4 to the data keys
of Precipitation values in the `data` object.