Add interactivity
Add id's to certain HTML elements to be able to refer to them from JavaScript.
Cache certain HTML element references on load to be able to refer to them
quickly throughout lifetime of the application.

Define `recalculate` function which takes the currently selected rain, land and
soil values, and uses the `data` object to lookup a `result` object, which has
an attribute each for Evapotranspiration `et`, Runoff `r`, and Infiltration
`i`. It then sets the text in the four labels, and adjusts height of the water
column.

Wire up `recalculate` to trigger on each tab shown (since the land and soil
types are implemented as bootstrap tabs), and slider change. Also trigger on
page load to initialize values.