Refactor getCookie method into helpers module
This method will be used by any code that queries user status. Since we
must query for already logged in users before the LoginModalView is
loaded, we must perform that query outside of this code block and must
have access to the `getCookie` method.

The CookieHelper class is named such to be descriptive in declaration,
but is exposed as `Cookie` for a cleaner API. `helpers.Cookie.get()` is
preferable to `helpers.CookieHelper.getCookie()`

Also fix typo in variable name, `crsftoken` -> `csrftoken`