Overload `/user/ajaxlogin` to get current user
Add additional behavior to `/user/ajaxlogin` endpoint that will supply
current user's credentials on GET if user is already logged in. If the
user is not logged in, no credentials are sent, but the server sends a
`guest` variable set to `true`.

Also hook HeaderView's `render` action to the `sync` event, so that
whenever we call `.fetch()` on the UserModel, the view is updated.