Initialize user if logged in, else show login modal
The new `getUserOrShowLogin` method attempts to use the newly overloaded
`/user/ajaxlogin` endpoint to fetch current user. If it succeeds, it
updates the user model, necessary to trigger a change and update the UI
since the view is listening for updates to the model. If the user is not
logged in and the server returns `guest` set to `true`, the login modal
is shown. If there is an error in communicating with the server, the
App.user has `guest` set to `true` by default, and thus the user will
still get a login modal.