Add user logout function to front-end
This function simply gets the `/user/ajaxlogout` URL created in the last
commit to log the user out. In addition to logging the user out, we may
also inspect the current route to see if this is a login-only view, in
which case we may redirect the user to the home page. This may require
refactoring this code elsewhere, since at the time of its definition it
doesn't really have access to the router.