Route logged out users to home
If users log out when they are in an area of the website that is for
logged in users only, then route them back to home. Else, leave them
where they are.

There are a number of problems with this commit, and I don't think it
should be a part of this pull request. Here are some of my objections:

 * We are adding more dependencies to App, such as the router which it
   did not require before and might not need for any other purpose.
 * We enumerate the "allowed" routes in an if condition. If we ever add
   new "allowed" routes in the future we will have to add them to here,
   which is not a clean solution.
 * Backbone.history.fragment() suffices for us now because there are no
   route parameters, but if we were to add them then it would no longer
   be enough.
 * Too much of user logic now resides in the App module, which I think
   should be encapsulated into its own module / controller / something.