Add AnsibleGalaxyHelper to remove noise from Vagrantfile
I broke out the Ansible Galaxy related cruft into a Ruby module that
gets required in the main `Vagrantfile`. The goal here is just to clean
things up before the fold so that people can get to the things they're
probably more interested in quicker.

Need to look into getting this into a Vagrant plugin or something at
some point.