Add `login` and `logout` methods to Backbone UserModel
These methods return a jqXHR object to which callbacks can be assigned
in the calling side, such as `.done()`, `.fail()`, or `.always()`. They
also encapsulate fundamental login and logout behavior, so that other
parts of the application only have to worry about the UI, not any of the
underlying network communication involved.

Also refactor LoginModalView to use UserModel's new `login` method in
place of doing a manual `user.fetch()`