Wait before ending bundle script
This fixes a problem where if you have multiple tasks running in
parallel, and the bundle script ends, all in progress tasks are killed.

For example, if you run "A & B" where A takes 1 second and B takes 10
seconds, then both will complete as expected. However, if you run "B &
A", then A will finish in 1 second, the script will exit, and B will be
killed prematurely.

This problem only happens if the last queued task finishes before
everything else.