Create functioning scenario tabs.
The scenario tabs need controls allowing them to be renamed, printed, deleted,
and shared. The names of scenarios need to be unique within a project. The
current scenario dropdown needs to trigger the appropriate tabs.

Removed the dummy scenarios and wired the UI to the models. Parent layout views
listen to events on sub-views to ensure that we check things like new scenario
names against the collection for uniqueness. This is also done to trigger the
changing of tabs from the current scenario dropdown.