Implement loading of modeling project from DB
If a project id is provided to the model URL,
load the corresponding modeling project.

- Generate client-side project models from server-side
API response.
- Add client-side and server-side routes to support
a model URL with a project ID parameter.
- Return a 404 if the project doesn't exist or the
logged in user isn't the owner of the project.
- Change attribute names to be snake case rather
than camel case so that they match Django model
attribute names. This allows us to avoid
maintaining a map between the front-end and
back-end models.

Refs #196