Add error controller and handlers
This ErrorController will handle all `/error[.*]` routes. Depending on
the route it will trigger a handler available in ErrorHandlers. If no
matching handler is found, it will trigger a generic handler.

For ITSI errors, this handler will show an error message in an alert
box, and show the login modal.