Implement saving of project from front-end.
Support saving changes made on the front-end to the
project and scenarios to the database as well as
creating new projects from a selected area of interest.

* Add a custom save method to the Project model to save
the project and all of the attached scenarios.
* Add a custom parse method to the Project model to create
a collection from the scenarios attribute.
* Create a new Project serializer that is used to update
an existing project. This serializer protects changes
to the User and Scenario objects by making them read-only.
* When loading a project, set project's area of interest
as the map's area of interest.
* Add djangorestframework_gis, which add serializers that
can convert between WKT > GeoJSON, and vice versa.
* Consolidate project init code in Project model init method.

Refs #196