Consolidate modal targets and fix cleanup.
The event listeners were not being properly removed when the modal was closed.
This adjusts that and adds modal close event listening which cleans up the modal
when clicking on a non-UI element (clicking off the modal closes it).
Additionally, use only one target for each kind of modal and reuse them.

Note that we cannot call view.destroy because that removes the view's el from
the DOM which prevents other modals from opening.