Use model cid instead of slug for toolbar tabs
Fixes a bug where you could not switch between tabs and also a bug
where the first tab was not initially active.

The problem is that Bootstrap does not automatically update the ID of
view container elements when the model is updated. Since the ID of the
tab view was based on the model slug, the tab ID would not update even
when the model slug had been changed.

The first solution I considered was to override the `render` function and
run `this.$el.attr('id', this.id())`, however, we would have to do this
everywhere throughout the app that uses a function to generate the view
ID.

I think it would be beneficial in the long run to rely more on model
cid fields rather than basing view IDs on model attributes.