Auto save project and scenarios.
We want the project to automatically save when a user has modified it either by
changing a property or by creating a new scenario. Scenarios and project details
can save independently to keep requests small and isoloated. Slightly more
advanced logic occurs on the first save because we need to associate the
scenarios with the newly created project id. Changes were also made to the REST
endpoints to avoid returning data that would overwrite the scenarios since they
are saved out of band.