Rerender the project page if the user changes.
We were encountering a timing issue in which the user fetch returned after the
project data causing the UI controls to be removed and preventing the user from
saving projects. Now, we listen to the App.user model and rerender the project
if it changes. This will also allow swapping to public mode if the user logs out
while on a public project.