Add CustomActionNode to create private DNS entry for cache
CloudFormation can create ElastiCache cache clusters, but it cannot
retrieve a specific cache cluster node's endpoint. To bridge the gap,
a `CustomActionNode` comes after the data plane stack to scoop up the
cache cluster ID, use it to determine the cache cluster node's endpoint,
and set it as a private `CNAME` record.

Connects #339