Add CustomActionNode to create a VPC endpoint for Amazon S3
By default, instances in the private subnet have to route traffic
through a NAT instance in a public subnet to communicate with Amazon S3.
Setting up a VPC endpoint allows the traffic to flow directly to Amazon
S3 from private subnets via a special route:

  https://aws.amazon.com/blogs/aws/new-vpc-endpoint-for-amazon-s3/

This is being done in a `CustomActionNode` because CloudFormation does
not have support for creating VPC endpoints. Also, many of the
additional Python classes are necessary because Boto does not have
direct support for describing and creating VPC endpoints.

Connects #287