Add application CloudFormation stack
The application servers live within an Auto Scaling Group (ASG) that is
associated with an Elastic Load Balancer (ELB). Health checks executed
by the ELB influence when the ASG adds or removes nodes from itself. All
application servers live inside private subnets and are exposed to the
internet via the ELB. In addition, SSL is terminated at the ELB.

Connects #298 and builds on top of #349 and #350.