Modify the router's execute method.
The router's `.execute` method has been modified so that explicit usage
of `router.navigate` in the `/model` and `/analyze` controllers works.
Make sure that the previous route is set correctly and that the execute
method returns `false` when arriving here via one of these explicit
calls.

Note that @casepyt is the author, not me.

Fixes #314