Add tiler CloudFormation stack
The tile servers live within an Auto Scaling Group (ASG) that is
associated with an Elastic Load Balancer (ELB). Health checks executed
by the ELB influence when the ASG adds or removes nodes from itself. All
tile servers live inside private subnets and are exposed to the internet
via the ELB.

Connects #297 and builds on top of #349, #350, #380.