Do not hard code aws_region.
We want to be flexible and allow for our applicaiton to be easily spun up in a
new region if necessary. To make this less problematic, remove hard coded
assumptions about the aws_region. This defaults the region to `None` unless
specifically passed to the email backend at instantiation. The boto mailer will
look up the instance region from the instance's metadata if it recieves no
aws_region upon its own instantiation.