Prevent double saving projects on initialization.
We were not able to restore saved projects because our initialization of the
project caused it to save not once but twice (when actually we did not want it
to save at all). This was caused by adding two scenarios to the new project.
Adding scenarios triggers an initial save. This is expected functionality however
adding two simultaneously caused two saves on the yet unsaved project which
caused two POST events rather than a POST and PUT. The scenarios were then
taking the last returned result and in most cases were getting the wrong project
id, which resulted in an error.

This prevents an initial save by adding the scenarios to the collection with the
silent option. Further, we introduce a flag on the project which is used to see
if a save has been attempted on the project and prevents saving if it has been
but an id has not yet been returned.