Add support for toggling DNS and removing stacks based on color
Adds a command line option to the `launch-stacks` subcommand that does
an `UPSERT` of the primary A record of `PublicHostedZoneName` using the
stack's load balancer endpoint.

If the `--activate-dns` flag is not provided, no DNS records are
altered.

Also, adds `remove-stacks` subcommand to remove all of the stacks
associated with a given color. The only stacks associated with colors
are:

- Application
- Tiler
- Worker

These changes also include a helper script that wraps all of the above
for CI, and a fix for a race condition during application and tiler
stack load balancer deletions.

Connects #464