Fix geocodeing issue.
It turns out that way back in PR #69 we forgot to add the key to our geocoder
endpoint. Python-Omgeo works by accepting query text and a key in the PlaceQuery
object. The `key` is passed on to the ESRI geocoder as magicKey. We had simply
omitted this and our results were thus not the result of doing a geocode on the
key but only on the query text. This resulted in ambiguous results.