Add Flower UI for Celery; use API endpoint for ELB health check
Installs and sets up Flower to provide a UI for the Celery workers. In
addition to the UI, Flower also provides API endpoints. These API
endpoints are exposed via Nginx, and pinged by a load balancer to
determine worker health. The following DNS records are also created:

- `blue-workers.mmw.azavea.com`
- `green-workers.mmw.azavea.com`

Their only purpose is to provide a way for humans to look at Flower
remotely. Access is scoped to the `IPAccess` configuration setting.

Connects #21 and #372