Don't fit bounds when toggling the model output
Restructure how views tell the map to go half size, so they can also
specify whether or not the bounds should be fit to the new size.  All
main region switching will still initially fit the AoI in the map, but
once the user begins toggling the model output pane, the bounds remain
consisten (and clipped if the map is shrinking).