Add application health check
Each probe initiates a connection attempt to the PostgreSQL database and
Redis. Failure of either returns an HTTP 503, while success of both
returns an HTTP 200. The health check is mounted at `/health-check`.

The inclusion of `BypassMiddleware` is needed because when the
connection to PostgreSQL is severed, other middleware end up triggering
exceptions before the `health_check` response can be returned. This
hurts our ability to know if the health check failure is just PostgreSQL
or PostgreSQL and Redis.

Connects #370