Add unit test for empty inputs
In this test we assert that an exception is raised and that it's message
is as expected when precipitation is not a specified input. We cannot
test the JSON output of the worker like we do for the success case in
`test_tr55_job_runs_in_chain`, because we set the `CELERY_ALWAYS_EAGER`
flag to True. This flag allows us to run these tests without having an
actual celery worker up, but unfortunately converts all `apply_async`
calls to `apply`, thus short-circuiting the error handling performed in
`core/tasks.py:save_job_error`.

If we want to test the JSON result, we will have to turn off the flag
`CELERY_ALWAYS_EAGER` and setup a mock worker for testing. There are
solutions such as https://github.com/RentMethod/celerytest which might
help. But that would be more along the lines of integration testing
rather than unit testing which is what we're doing here. See discussion
on http://stackoverflow.com/a/17155922/2053314 for details.