REST service.
A spray-based REST service has been created.  The user POSTs a polygon
stored as a GeoJson file, and receives a count of land-use, soil-type
pairs in the intersection of the query polygon and the test area.

Connects #545

The goals discussed in #544 can be done by sending a GeometryCollection
instead of a Polygon or MultiPolygon.  Those results can then be woven
into a census of the type expected by the TR-55 code.

Connects #544