Setting to split worker queues for deployments.
When doing blue/green deployments we can end up in a situation where there are
two stacks, with works and apps both attached to the same redis instance which
acts as the celery broker. This means that new jobs may be picked up by the old
workers which is likey to break things. Here we provide environment variables
for the stack colors to the worker and app instances and use these as the queue
names to prevent the wrong stack from running jobs.