Return to original URL after ITSI login
Logging in with ITSI requires us to do an OAuth dance where we redirect
a few times. By carrying the original request path in a `next` query
string parameter, we ensure that once that dance is finished the user is
returned to their originally requested path.

If at any stage the `next` query string parameter is not specified, then
we redirect to `/`.

We also use this new architecture to redirect users logging in via ITSI
from the UI. Thus, if a user is viewing a public project, and then
clicks "Login" and then "ITSI", after a successful login they will be
redirected back to the project they were viewing.