Initialize & save project if query param is set.
In some cases we want a project to be saved as soon as the user lands on the
site. This is determined by the addition of a URL query parameter. If this
parameter is set, a project will be instantly created and saved. If the user
designates an area of interest and then models it everything will be saved. If
they then move back out and redraw, the same project will be used and the data
will be overwritten.