Rework settings module.
I was having trouble keeping track of the activityMode flag without creating
circular dependencies on app. The global setttings config seemed the best
place to handle this but the getter and setter were implemented infeasable to
use. This reworks the settings module and updates the existing code to use the
new methods. The activityMode flag is then stored and retreived via settings.