Always stop editing on blur
We set the `contenteditable` attribute to `true` as part of this
`renameScenario` method, and it is removed whenever there is a call to
`this.render()` since the template itself doesn't have that set.

However, `this.render()` is only called if there are any changes to the
model. Thus, in cases when a user starts renaming a scenario but then
clicks outside without changing the value, `this.render()` was never
fired and the editing would not stop.

By replicating the logic in the `keypress` event and explicitly calling
`this.render()` in case of no changes we ensure that editing is stopped
even when there are no changes to the name.

Since the logic is now replicated in both `keypress` and `blur` events,
we extract it into a `setScenarioName` function that is used in both
places.

To reuse logic already available in `updateScenarioName`, we make that
function always return a value. If the value is truthy, then the change
was successful and `this.render()` was called automatically, otherwise
we call it manually.

We also add a user-facing alert informing the user that they chose a
name that is already in use, and that they must choose unique names for
scenarios.