Preserve original ordering of Collections
For the purpose of producing consistent MD5 hashes for caching, #567
added code for keeping many Collections in an internally sorted order
in 220a6e8. The actual order (usually the order in which they were added
to the collection) was not meaningful at the time.

Now, however, we are solving problems which do require us to respect the
order in which models are added to the collection, e.g. #641. Thus, we
undo automatic sorting and also remove tests for it. We adjust some more
tests to respect the order of models within a collection, but still test
that models with differently sorted keys are still evaluated the same.

Since Collections are stringified to Arrays and not Objects in JSON, the
sorting there should be preserved automatically, without us having to
sort each Collection. Thus, this change should not affect the stability
of our caching approach.