Draw one Km square on click.
When the user select the one square km drawing tool, they are given a point to
drop on the map. This acts as the center of the square. We use turf
destination, which uses the Haversine formula to calculate four points around
the center at a distance of one half Km with the bearings -180, -90, 0, 90.
This is then used to create a polygon and start the analysis.