Load previously saved state
If the `interactiveState` has been set at any point in the past, we may
receive it via the `loadInteractive` postMessage. We check if this URL
differs from the current one, and if so we check if a new project has
been created for this user (since we are in the embed mode). If so, we
delete the existing project. Finally, we redirect to the specified URL.