Allow ITSI users to clone projects via UI
Create a new "Clone Project" link in the user drop-down for ITSI users
that allows them to specify a project ID and clone it.

This is required because users won't be able to paste in a URL when the
app is embedded inside an activity.

Before attempting to redirect to the clone URL, we try to do a fetch of
the given project ID. If the user does not have read access to it, we
show them an alert message instead.