Make modeling results return metric numbers.
Modeling results have been returned in Imperial units and the coverage area
was returned in the number of pixels in the analyzed raster. This is not a
useful number for most users. We convert the results to metric units prior to
returning the results. Specifically, the coverage requires information about
the analyzed area so that it can roughly calculate the area of a pixel based
on the latitude.