Save the AOI name with the project.
We had not been saving the AOI name when saving the project. As a result, if
you reloaded a project and then went back to analyze mode, the name would be
blank. This change allows the AOI name to be persisted as part of the modeling
project and then restores it to the map when the project is reloaded.