Update project, state, when AoI changes
When in Activity Mode, update the project on AoI change. When in Embed
Mode, communicate this to the parent container via the ITSI interface.

Only do this when there are no existing scenarios. Otherwise we can find
ourselves in a spot where the AoI has changed and saved but the old
scenarios still exist with the old modifications which are no longer
valid.

When there are existing scenarios, the proper deletion and save will be
performed when moving from /analyze to /project.