Don't save empty projects in embed mode
Previously we would save a new project as soon as the app initialized in
embed mode. This had the effect of creating a number of empty projects
when the app was loading an existing project. We would delete this empty
project before loading the specified one.

Now, we still create the project as soon as the app starts, but defer
saving it to the database until an Area of Interest has been drawn and
the user lands in the Analyze view. This ensures that the project is
always saved as soon as an AoI is available, which is fine since we're
not tracking any other changes that we could restore (such as the maps
zoom or location, or values in the search box).

This further simplifies the loading logic, since now we don't have to
check for and delete an existing project. We simply reset `currProject`
and navigate to given route, and let the controllers handle the rest.