Ensure "Current Conditions" is first scenario
In some edge cases, the POST requests for non-Current Conditions
scenarios would be processed before Current Conditions, resulting in
them having a lower `id`, and thus showing up before Current Conditions
in the UI.

When loading a new set of scenarios, we check to see if the first one
is Current Conditions. If not, we find it and move it to first place.

Also add a test to exercise this new behavior.