Fix Safari infinte location request loop.
In some cases, it seems requesting geolocation at page load will cause Safari
to continue to ask for location permission in an infinite loop.
https://stackoverflow.com/questions/27150465/
The solution to this seems to be triggering the location request based on some
user event or supplying a short delay to allow whatever is causing Safari the
issue time to resolve itself.