Implement support for Google dynamic max zoom
When using Google satellite layers, the max zoom level changes
based on the available imagery.

* Based on the location of the user, update max zoom level for
Google layers.
* Because of the way the max zoom level is set (it's read from the
layer options when a layer is selected in the layer control), changes
to the max zoom level do not take effect until the user switches
to another layer, and then back to a Google layer. As a stop gap,
we get the max zoom level of the initial map location and update the
setting for a Google layer immediately. We also update it as the user
moves the map regardless if a Google layer is active. While a Google
layer is active, if the user happens to pan to an area that does not
support the current zoom level, the user should see "Sorry, we have no
imagery here" tiles. This is a graceful fallback until a smoother
integration is possible; which will be available in Leaflet 1.0. This
version has better support for dynamically updating the max zoom level
of the map.

Refs #692