Allow boundary dropdown to be clicked.
Boundary layers were being triggered based on the data attribute of the
clicked element in the dropdown. But since spans were added to the text for
styling this broke the existing functionality. By changing the click handler
to work with the currentTarget (the specific element the event is bound to) we
will always capture the proper element with the data attribute in question.
Also added bounding to the popovers to ensure they stay within the bounds of
the main area.