Save scenarios silently
We trigger scenario saves when there are changes to the scenario name or
modifications or inputs, so as to facilitate auto-saving and relieve the
user of the responsibility to save manually. Unfortunately, this call to
`save` would trigger a call to `parse` which would update the collection
attributes with new values received from server, thus triggering many
views to refresh.

We only want to update views when loading new values, thus when `parse`
is called after a `fetch`. To ensure that `parse` does not affect values
after a `save`, we pass in a `silent` option which is inspected within
`parse`, and if found we simply return existing values without changing
any, which does not trigger a view refresh.

Individual views track specific fields on the scenario, thus they update
correctly, and in isolation.

Refs #732
Refs #672