Add support for SJS; load it with geoprocessing JAR
Spark Job Server (SJS) has been added to the `worker` virtual machine
within a Docker container. Upstart manages the SJS container, so all
normal interactions with it should be via the `start`, `stop`, and
`restart` commands.

Once SJS is up-and-running, a second Upstart job fires to load the
geoprocessing JAR. The `geoprocessing` Spark context is created by SJS
itself. Any additional tuning we do to the context should happen via the
SJS configuration file.

Amazon credentials need to be mapped into the `worker` virtual machine
for anything to work (the geoprocessing process makes tile requests to
Amazon S3). This happens via a Vagrant mount for the `vagrant` user, but
also via an Ansible `copy` task for the geoprocessing service itself.

Right now, SJS jobs are synchronous, but I believe that they'll need to
be made asynchronous (with the Celery job polling for status), or else
we are going to run into issues with long-running synchronous SJS jobs.

I also increased the JavaScript polling timeout from 5 seconds to 10.
This allows more successfully completed SJS jobs to have their results
make it back to the UI before giving up on polling.

Attempts to resolve #50.