After SJS HTTP socket binds, wait an additional second
We were attempting to `POST` the SJS job JAR too quickly after the SJS
service starts. This adds a one second sleep before we attempt the first
`POST` (in addition to the `nc` in a loop at the end of the SJS service
definition waiting for a successful connection to port 8090.