Update ITSI with URL of new project upon creation
Previously, the project ID would be available but not present in the
URL, resulting in a use case when a student who has progressed to the
project stage closes the window and returns to the draw stage, since the
project URL was never updated / communicated to the containing activity.
By ensuring that the project URL is updated as soon as the project stage
is entered, we avoid that use case.

Refs #804