Add debug level logging to Spark Job Server
Adds a Log4j configuration file setup to override the one provided by
default in the Spark Job Server container. In development, the overall
log level is set to `DEBUG` and the AWS SDK log levels to `DEBUG`. In an
AWS deployment, the overall log level is set to `INFO` and the AWS SDK
log levels to `WARN`.

Useful information provided:

- Log entries of actual tile URLs being pulled down by GeoTrellis
- Log entries prefixed with timestamps
- Timestamps in logs align in Logstash