Show 100% Forest Cover scenario in Compare View
Since the Compare View needs scenarios, and 100% Forest Cover data is
contained inside the Current Conditions scenario, we make a copy of it
with is_current_conditions set to false and is_pre_columbian set to true
before passing it to the Compare View.

When rendering TR55 Runoff views, we check if the current scenario is
Current Conditions, Pre Columbian, or a regular one in Compare View. We
use unmodified, pc_unmodified, or modified results respectively.