Adjust layer control so that it is never obscured.
Previously, the element that contained the layer list had a parent with
a very small size and a very deep placement within the map container
hierarchy. To better size and position the list, we make it a direct
sibling of the map container, which is the controlling factor for the
size of the list.

* Refactor layer control code to create two separate elements, one for
the control button and one with the layer list.
* Adjust style of layer list so that it grows and shrinks along with the
map height, is positioned near the top of the map so it has more
vertical space, and so that any hidden layers are accessible by a
scroll bar.
* Adjust style of draw tools container so that it does not overlap the
layer list on small screens.

Refs #843