Allow use of back and forward buttons
This attempts to fix a bug that prevents navigation form the modeling
page, to the analyze page, then back to the modeling page via the
browser back and forward buttons.

The problem was being caused by the fact that the
'#modeling-details-region' element did not come into existence early
enough when you visited the modeling view with browser forward button
compared to when you did with the UI button.  A lock has been introduced
to make sure that that element exists before it is used.

Connects #725